feat: add helpers
